requirements.txt
PyQt6>=6.4.0
requests>=2.28.0
humanize>=4.0.0
python-libtorrent>=2.0.0
win10toast>=0.9.0 